+++ src/classes/queue-getters.ts
@@ -91,13 +91,36 @@ export class QueueGetters<
   /**
     Returns the number of jobs waiting to be processed. This includes jobs that are "waiting" or "delayed".
   */
-  count(): Promise<number> {
-    return this.getJobCountByTypes(
+  async count(): Promise<number> {
+    const count = await this.getJobCountByTypes(
       'waiting',
       'paused',
       'delayed',
       'waiting-children',
     );
+
+    if (await this.hasDelayedMarker('wait', 'paused')) {
+      return count - 1;
+    }
+
+    return count;
+  }
+
+  private async hasDelayedMarker(
+    ...types: ('wait' | 'paused')[]
+  ): Promise<boolean> {
+    const client = await this.client;
+
+    const promises = [];
+    if (types.includes('wait')) {
+      promises.push(client.lindex(this.toKey('wait'), 0));
+    }
+
+    if (types.includes('paused')) {
+      promises.push(client.lindex(this.toKey('paused'), 0));
+    }
+
+    return (await Promise.all(promises)).some(value => value === '0');
   }
 
   /**
@@ -135,6 +158,19 @@ export class QueueGetters<
     res.forEach((res, index) => {
       counts[currentTypes[index]] = res[1] || 0;
     });
+
+    if (counts['wait'] > 0) {
+      if (await this.hasDelayedMarker('wait')) {
+        counts['wait']--;
+      }
+    }
+
+    if (counts['paused'] > 0) {
+      if (await this.hasDelayedMarker('paused')) {
+        counts['paused']--;
+      }
+    }
+
     return counts;
   }