+++ tests/test_sandboxed_process.ts
@@ -7,7 +7,6 @@ import {
   Job,
   Queue,
   QueueEvents,
-  QueueScheduler,
   Worker,
 } from '../src/classes';
 import { beforeEach } from 'mocha';
@@ -177,9 +176,6 @@ describe('sandboxed process', () => {
     it('moves job to failed', async function () {
       this.timeout(6000);
 
-      const queueScheduler = new QueueScheduler(queueName, { connection });
-      await queueScheduler.waitUntilReady();
-
       const processFile =
         __dirname + '/fixtures/fixture_processor_unrecoverable.js';
 
@@ -219,7 +215,6 @@ describe('sandboxed process', () => {
       expect(state).to.be.equal('failed');
 
       await worker.close();
-      await queueScheduler.close();
     });
   });
 
@@ -336,7 +331,8 @@ describe('sandboxed process', () => {
   });
 
   it('should process and update progress', async () => {
-    const processFile = __dirname + '/fixtures/fixture_processor_progress.js';
+    const processFile =
+      __dirname + '/fixtures/fixture_processor_update_progress.js';
 
     const worker = new Worker(queueName, processFile, {
       connection,
@@ -354,11 +350,6 @@ describe('sandboxed process', () => {
           expect(progresses).to.be.eql([10, 27, 78, 100]);
           expect(Object.keys(worker['childPool'].retained)).to.have.lengthOf(0);
           expect(worker['childPool'].getAllFree()).to.have.lengthOf(1);
-          const logs = await queue.getJobLogs(job.id);
-          expect(logs).to.be.eql({
-            logs: ['10', '27', '78', '100'],
-            count: 4,
-          });
           resolve();
         } catch (err) {
           reject(err);
@@ -371,6 +362,34 @@ describe('sandboxed process', () => {
     });
 
     await queue.add('test', { foo: 'bar' });
+
+    await completing;
+    await worker.close();
+  });
+
+  it('should process and update data', async () => {
+    const processFile = __dirname + '/fixtures/fixture_processor_update.js';
+
+    const worker = new Worker(queueName, processFile, {
+      connection,
+      drainDelay: 1,
+    });
+
+    const completing = new Promise<void>((resolve, reject) => {
+      worker.on('completed', async (job: Job, value: any) => {
+        try {
+          expect(job.data).to.be.eql({ foo: 'baz' });
+          expect(value).to.be.eql('result');
+          expect(Object.keys(worker['childPool'].retained)).to.have.lengthOf(0);
+          expect(worker['childPool'].getAllFree()).to.have.lengthOf(1);
+          resolve();
+        } catch (err) {
+          reject(err);
+        }
+      });
+    });
+
+    await queue.add('test', { bar: 'foo' });
 
     await completing;
     await worker.close();